Double

+++ main.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int main()
 {
 cout << "Enter A and B: ";
-float a, b;
+double a, b;
 cin >> a >> b;
 cout << "A + B = " << a + b << '\n' << "A - B = " << a - b << '\n' << "A * B =" << a * b << '\n'
 << "A / b = " << a / b * 1.0 << '\n';
